requirements.txt
langgraph
langchain-groq
langchain-core
pymongo
tavily-python
praw
requests
beautifulsoup4
python-dotenv
pydantic
fastapi
uvicorn
sse-starlette